decouple deployment configuration from legacy migration and swing dependency
